requirement.txt
albumentations==1.4.4
opencv-python==4.9.0.80
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
PyYAML==6.0.1
